Moved Components to separate file, moved logic to ViewModel
