+++ numeratio/View/ContentView.swift
@@ -40,7 +40,7 @@ struct ContentView: View {
                     .disabled(true)
                     .opacity(0.5)
                   
-                    NavigationLink (destination: GameHistory()) {
+                    NavigationLink (destination: GameHistoryView()) {
                         MenuItem(
                             title: "History",
                             caption: "Analyze your counting speed history",
@@ -58,66 +58,7 @@ struct ContentView: View {
     }
 }
 
-struct Background: View {
-    var body: some View {
-        GeometryReader{ geometry in
-            Circle()
-                .size(width: geometry.size.width * 2, height: geometry.size.width * 2)
-                .fill(Color("BackgroundColor"))
-                .position(y: -(geometry.size.height / 2.5))
-        }
-    }
-}
 
-struct Logo: View {
-    @Environment(\.colorScheme) private var colorScheme
-    
-    var body: some View {
-        Image(colorScheme == .dark ? "LogoDark" : "LogoLight")
-            .resizable()
-            .frame(width: 160, height: 160)
-            .padding()
-            .padding(.top, 20)
-    }
-}
-
-struct MenuItem: View {
-    let title: String
-    let caption: String
-    let image: String
-    
-    var body: some View {
-        HStack {
-            Image(systemName: image)
-                .resizable()
-                .aspectRatio(contentMode: .fit)
-                .frame(width: 32)
-                .padding(.trailing, 8)
-                .foregroundStyle(Color.black)
-            
-            VStack {
-                HStack {
-                    Text(title)
-                        .foregroundStyle(Color.black)
-                        .font(.title2)
-                        .fontWeight(.bold)
-                    Spacer()
-                }
-                HStack {
-                    Text(caption)
-                        .foregroundStyle(Color.black)
-                        .font(.caption)
-                    Spacer()
-                }
-            }
-            Spacer()
-        }
-        .frame(width: 320, height: 40)
-        .padding()
-        .background(Color("InteractiveColor"))
-        .cornerRadius(8)
-    }
-}
 
 #Preview {
     ContentView()
